Service Fee Tests

+++ src/main/java/Card/AmericanExpress.java
@@ -6,7 +6,7 @@ public class AmericanExpress extends Card {
     }
 
     @Override
-    public int serviceFee() {
+    public float serviceFee() {
         return 0;
     }
